LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-getAttributeType-svg-expected.txt


PASS trustedTypes.getAttributeType html script[href]
FAIL trustedTypes.getAttributeType svg script[href] assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL trustedTypes.getAttributeType svg script[href] xlink href assert_equals: expected (string) "TrustedScriptURL" but got (object) null
PASS trustedTypes.getAttributeType svg script[href] other href
